refactor resources mod
