+++ hermes/bin/src/packaging/resources/bytes.rs
