+++ hermes/bin/src/packaging/resources/fs.rs
@@ -20,6 +20,11 @@ impl FsResource {
         Self(path.as_ref().to_path_buf())
     }
 
+    /// Get resource path.
+    pub(crate) fn get_path(&self) -> PathBuf {
+        self.0.clone()
+    }
+
     /// Update current resource to make it relative to the given path.
     pub(crate) fn make_relative_to<P: AsRef<Path>>(&mut self, to: P) {
         if self.0.is_relative() {
